gpu-containers/requirements.txt
docker-py
tabulate